refactor: move file
Move error-logging module into cli folder, since it is concerned with cli
